refactor: move service logic to service
